Added trash that spawns off screen and move randomly
Added trash objects the spawn 12 units to the left of the player. They move to the right in a random vertical direction. They move in both directions at random speeds. They don't collide with the invisible walls around the tiles.

In the future, the player will be able to pick up trash as items to craft with. I would also like them to collide with the tile sprites instead of phasing through them.